Revert unrelated test-runner refactors and align the Durable Workflow form title
- runAiTest.ts and runLibsIntegrationTest.ts are restored to their upstream
  state: their migration to the shared runTests helper was general cleanup
  not needed by this PR (the workflow snapshot tests run through the main
  runTest.ts entrypoint only). The runner changes that the new tests do
  need are kept: the main runner's env-webpack-helper path fix, the .md
  require hook and TEST_FILE_GREP filter in test/index.ts, and the shared
  helper installing extension dependencies into the launched extensions
  dir so the extension activates in the test host.

- FunctionForm: formType now reads 'Durable Workflow' so the form and page
  titles match the artifact rename used elsewhere (getFunctionType already
  did), with the title subtitle updated accordingly.
