+++ packages/ballerina-extension/test/runAiTest.ts
@@ -17,9 +17,11 @@
  */
 
 import * as path from 'path';
-import { runTests } from './lib/index';
+import * as cp from 'child_process';
+import { downloadAndUnzipVSCode, resolveCliPathFromVSCodeExecutablePath, runTests } from '@vscode/test-electron';
 const dotenv = require('dotenv');
-const { createEnvDefinePlugin } = require('../../../../common/scripts/env-webpack-helper');
+const packageJson = require('../../package.json');
+const { createEnvDefinePlugin } = require('../../../../../common/scripts/env-webpack-helper');
 
 async function main() {
     try {
@@ -40,9 +42,28 @@ async function main() {
             );
         }
 
-        // Run tests with specific grep pattern for AI tests. The shared runTests helper downloads
-        // VS Code and installs the extension dependencies into the launched extensions dir
+        // Download VS Code and get CLI path
+        const vscodeExecutablePath = await downloadAndUnzipVSCode();
+        const [cli, ...args] = resolveCliPathFromVSCodeExecutablePath(vscodeExecutablePath);
+
+        // Install extension dependencies if they exist
+        if (packageJson.extensionDependencies && packageJson.extensionDependencies.length > 0) {
+            console.log(`Installing ${packageJson.extensionDependencies.length} extension dependencies...`);
+            for (const extensionId of packageJson.extensionDependencies) {
+                console.log(`Installing extension: ${extensionId}`);
+                cp.spawnSync(cli, [...args, '--install-extension', extensionId], {
+                    encoding: 'utf-8',
+                    stdio: 'inherit',
+                });
+            }
+            console.log('Extension dependencies installed successfully');
+        } else {
+            console.log('No extension dependencies to install');
+        }
+
+        // Run tests with specific grep pattern for AI tests
         await runTests({
+            vscodeExecutablePath,
             extensionDevelopmentPath,
             extensionTestsPath,
             launchArgs: [
